+++ nf_core/components/install.py
@@ -16,6 +16,7 @@ from nf_core.components.components_command import ComponentCommand
 from nf_core.components.components_utils import (
     get_components_to_install,
     prompt_component_version_sha,
+    try_generate_container_configs,
 )
 from nf_core.components.constants import (
     NF_CORE_MODULES_NAME,
@@ -157,7 +158,7 @@ class ComponentInstall(ComponentCommand):
         if not self.install_component_files(component, version, self.modules_repo, install_folder):
             return False
 
-        # Update module.json with newly installed subworkflow
+        # Update module.json with newly installed component
         modules_json.load()
         modules_json.update(
             self.component_type, self.modules_repo, component, version, self.installed_by, install_track
@@ -166,6 +167,10 @@ class ComponentInstall(ComponentCommand):
         if self.component_type == "subworkflows":
             # Install included modules and subworkflows
             self.install_included_components(component_dir)
+
+        # Regenerate container configuration files for the pipeline when modules are installed
+        if self.component_type == "modules":
+            try_generate_container_configs(self.directory, self.modules_repo.repo_path)
 
         if not silent:
             modules_json.load()