+++ nf_core/components/remove.py
@@ -8,6 +8,7 @@ from rich.syntax import Syntax
 
 import nf_core.utils
 from nf_core.components.components_command import ComponentCommand
+from nf_core.components.components_utils import try_generate_container_configs
 from nf_core.modules.modules_json import ModulesJson
 
 from .install import ComponentInstall
@@ -172,6 +173,10 @@ class ComponentRemove(ComponentCommand):
                     # remember removed dependencies
                     if dependency_removed:
                         removed_components.append(component_name.replace("/", "_"))
+            # Regenerate container configuration files for the pipeline when modules are removed
+            if self.component_type == "modules":
+                try_generate_container_configs(self.directory, repo_path)
+
             # print removed dependencies
             if removed_components:
                 log.info(f"Removed files for '{component}' and its dependencies '{', '.join(removed_components)}'.")